upd PresetListWidget

+++ components/views/presetsList.php
@@ -8,6 +8,7 @@
 use yii\helpers\Url;
 use yii\helpers\Html;
 use yii\widgets\ActiveForm;
+use app\components\PresetsListWidget;
 
 ?>
 <?php if ($presets): ?>
@@ -22,16 +23,28 @@ use yii\widgets\ActiveForm;
                         </h5>
 <!--                        <p class="card-text text-muted mb-0">Теги или краткое описание, дата, время и т.д.</p>-->
                         <?php if   ($preset->discipline): ?>
-                        <span class="badge badge-info dropdown-toggle" id="navbarDropdownMenuLink" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Подробнее</span>
-                        <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
+
+                        <span class="badge badge-info dropdown-toggle" data-toggle="collapse" data-target="#collapseDiscipline<?= $preset->id ?>" aria-expanded="false" aria-controls="collapseDiscipline<?= $preset->id ?>">
+                            Подробнее
+                        </span>
+
+                        <div class="collapse" id="collapseDiscipline<?= $preset->id ?>">
                             <?php foreach ($preset->discipline as $discipline): ?>
-                            <span class="dropdown-item"><?= $discipline->name ?></span>
+                                <p class="card-text text-muted mb-0"><?= $discipline->name ?></p>
                             <?php endforeach; ?>
+
+                            <span class="badge badge-secondary" data-toggle="collapse" data-target="#collapseDiscipline<?= $preset->id ?>" aria-expanded="false" aria-controls="collapseDiscipline<?= $preset->id ?>">
+                            Скрыть
+                        </span>
                         </div>
+
                         <?php else: ?>
                              <p class="card-text text-muted mb-0">Нет упражнений</p>
                         <?php endif; ?>
                     </div>
+
+                    <?php if ($purpose == PresetsListWidget::PURPOSE_EDIT): ?>
+
                         <?php if ($preset->user_id === Yii::$app->user->id): ?>
                             <?php $form = ActiveForm::begin(['action' => Url::to(['/preset/delete']), 'method' => 'delete']) ?>
                             <?= Html::hiddenInput('Presets[id]', $preset->id)?>
@@ -65,11 +78,30 @@ use yii\widgets\ActiveForm;
                                         ]) ?>
                             </div>
                         <?php endif; ?>
+
+
+                    <?php elseif ($purpose == PresetsListWidget::PURPOSE_START): ?>
+
+                        <div class="row ml-auto">
+                            <?= Html::a(
+                                Html::button('<i class="material-icons">play_circle_filled</i>',
+                                    [
+                                        'class' => 'nav-link-icon mr-2 mb-2 p-1 btn'
+                                    ]),
+                                Url::to(['/set/start', 'id' => $preset->id]),
+                                [
+                                    'class' => 'ml-auto'
+                                ]
+                            ) ?>
+                        </div>
+
+                    <?php endif; ?>
+
                 </div>
             </div>
         </div>
     </div>
 <?php endforeach; ?>
 <?php else: ?>
-<h5>У вас нет программ</h5>
+<h5>У тебя нет программ для выбора. Создай свою первую программу и начни тренировку...</h5>
 <?php endif; ?>
